+++ plugins-bundle-analyzer/runAnalyzeBundles.ts
@@ -1,0 +1,3 @@
+import { analyze } from './analyzeBundles';
+
+analyze();