+++ plugins-bundle-analyzer/src/ricos-common.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { RicosEngine } from 'ricos-common';
-
-export default () => {
-  return <RicosEngine />;
-};