+++ plugins-bundle-analyzer/src/ricos-common.tsx
@@ -1,0 +1,12 @@
+import React, { FunctionComponent } from 'react';
+import { RicosEngine } from 'ricos-common';
+
+const functionComponent: FunctionComponent = () => <div />;
+
+export default () => {
+  return (
+    <RicosEngine RicosModal={functionComponent} isViewer>
+      <div />
+    </RicosEngine>
+  );
+};