+++ scripts/templates/atomic-plugin/web/src/toolbar/createToolbar.ts
@@ -2,12 +2,12 @@ import createInlineButtons from './inline-buttons';
 import createInsertButtons from './insert-buttons';
 import { CreatePluginToolbar } from 'wix-rich-content-common';
 
-const createToolbar: CreatePluginToolbar = ({ helpers, t, settings, isMobile }) {
+const createToolbar: CreatePluginToolbar = ({ helpers, t, settings, isMobile }) => {
   return {
     InlineButtons: createInlineButtons({ t, settings, isMobile }),
     InsertButtons: createInsertButtons({ helpers, t, settings, isMobile }),
     name: 'yourPluginName',
   };
-}
+};
 
 export default createToolbar;