+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 - `ricos-editor`
   - [#1300](https://github.com/wix-incubator/rich-content/pull/1300) missing imported styles
   - [#1296](https://github.com/wix-incubator/rich-content/pull/1296) Editor Modal z-index increase
+### :house: Internal
+- `plugins-bundle-analyzer`
+  - [#1302](https://github.com/wix-incubator/rich-content/pull/1302) converted analyzer to typescript
   - [#1303](https://github.com/wix-incubator/rich-content/pull/1303) actionColor fix for "more+" button
   - [#1304](https://github.com/wix-incubator/rich-content/pull/1304) textColor fix for plugin (+) button
   - [#1305](https://github.com/wix-incubator/rich-content/pull/1305) theme - toolbar white on white